refactoring and todo
