+++ src/main/java/graphics/playPage/PlayPage.java
@@ -19,7 +19,7 @@ import java.awt.event.ActionListener;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-//TODO: add javadoc
+//TODO: add javadocs
 public class PlayPage extends CommonPage {
     private Observer observer;
