+++ src/main/java/graphics/playPage/leftPanel/LeftPanel.java
@@ -1,19 +1,22 @@
 package graphics.playPage.leftPanel;
 
 import graphics.Window;
-import graphics.playPage.leftPanel.sliders.InertiaFactorSlider;
 import graphics.playPage.observer.Observer;
+import graphics.playPage.leftPanel.labelSection.LabelSection;
+import graphics.playPage.leftPanel.sliders.InertiaFactorSlider;
+import graphics.playPage.leftPanel.leftDrawPanel.LeftDrawPanel;
 
 import graphicsEngine.colors.SimpleColorScheme;
 import graphicsEngine.presets.panels.VerticalPanel;
 
-import java.awt.*;
-import javax.swing.*;
+import java.awt.Color;
+import java.awt.LayoutManager;
+import javax.swing.BoxLayout;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-//
+//TODO: add javadocs
 public class LeftPanel extends VerticalPanel {
     private static final int PANEL_WIDTH = 150;
     private static final int BACKGROUND_BRIGHTNESS = 80;
@@ -24,7 +27,7 @@ public class LeftPanel extends VerticalPanel {
                     BACKGROUND_BRIGHTNESS),
             TEXT_COLOR = null; // default - white
 
-    //
+    //TODO: add javadoc
     public LeftPanel(@NotNull Window window,
                      @NotNull Observer observer) {
         super(
