+++ src/main/java/graphics/playPage/leftPanel/leftDrawPanel/AbstractTextSection.java
@@ -1,22 +1,22 @@
-package graphics.playPage.leftPanel;
+package graphics.playPage.leftPanel.leftDrawPanel;
 
+import java.util.Objects;
 import java.awt.Graphics;
 import java.awt.Color;
-import java.util.Objects;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-//
+//TODO: add javadocs
 abstract class AbstractTextSection extends AbstractSection {
     private int currentLine;
 
-    //
+    //TODO: add javadoc
     AbstractTextSection() {
         super();
     }
 
-    //
+    //TODO: add javadoc
     @Override
     void draw(@NotNull Graphics g, @NotNull Color textColor,
               int verticalOffset) {
@@ -24,7 +24,7 @@ abstract class AbstractTextSection extends AbstractSection {
         currentLine = 0;
     }
 
-    //
+    //TODO: add javadoc
     final void drawLineOfText(@NotNull Graphics g, @NotNull Color textColor,
                               int @NotNull [] sectionLocation,
                               @Nullable String text) {
@@ -35,13 +35,13 @@ abstract class AbstractTextSection extends AbstractSection {
                 sectionLocation[1] + currentLine * textSpacing);
     }
 
-    //
+    //TODO: add javadoc
     @Override
     final int getSectionHeight() {
         return getLastLineIndex() * textSpacing;
     }
 
-    //
+    //TODO: add javadoc
     int getLastLineIndex() {
         return currentLine;
     }
