+++ src/main/java/graphics/playPage/leftPanel/leftDrawPanel/FieldInfoSection.java
@@ -1,4 +1,4 @@
-package graphics.playPage.leftPanel;
+package graphics.playPage.leftPanel.leftDrawPanel;
 
 import staticData.StaticData;
 
@@ -9,14 +9,15 @@ import consoleUtils.NumberFormatter;
 
 import org.jetbrains.annotations.NotNull;
 
-//
+//TODO: add javadocs
 class FieldInfoSection extends AbstractTextSection {
-    //
+
+    //TODO: add javadoc
     FieldInfoSection() {
         super();
     }
 
-    //
+    //TODO: add javadoc
     @Override
     final void draw(@NotNull Graphics g,
               @NotNull Color textColor,
